move AAC-helper to remux components to
